+++ app/Apis/Summ_Ai/Summ_Ai.php
@@ -819,7 +819,6 @@ class Summ_Ai extends Base implements Api_Base {
 		// add setting.
 		$setting = $settings_obj->add_setting( 'easy_language_summ_ai_source_languages' );
 		$setting->set_section( $summ_ai_tab_main );
-		$setting->set_show_in_rest( true );
 		$setting->set_type( 'array' );
 		$setting->set_default( $languages );
 		$field = new Checkboxes();
@@ -846,7 +845,6 @@ class Summ_Ai extends Base implements Api_Base {
 		// add setting.
 		$setting = $settings_obj->add_setting( 'easy_language_summ_ai_target_languages' );
 		$setting->set_section( $summ_ai_tab_main );
-		$setting->set_show_in_rest( true );
 		$setting->set_type( 'array' );
 		$setting->set_default( $languages );
 		$field = new FieldTable();