+++ app/Plugin/Languages.php
@@ -45,7 +45,7 @@ class Languages {
 	}
 
 	/**
-	 * Get actual supported languages.
+	 * Return actual supported languages.
 	 *
 	 * If an API is active, use their supported languages if they can be configured.
 	 * If no API is active, use plugin settings.