+++ changelog.md
@@ -5,7 +5,9 @@
 ### Added
 
 - Revamped plugin for a more robust and sustainable structure
+- Added a new object for handle all plugin settings
 - Added support for the PageBuilders Beaver Builder, Bold Page Builder, Brizy, Kubio, SiteOrigin, SeedProd and Visual Composer
+- Added option to reset the plugin in backend settings (in preparation for Cyber Resilience Act)
 
 ### Changed